# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pilot Web UI API",
    "description": "BFF server API for Pilot Web UI — chat sessions, platform channels, model management, skills, memory, logs, and terminal.",
    "version": "0.2.6"
  },
  "servers": [
    { "url": "http://localhost:8648", "description": "Local development" }
  ],
...
  "paths": {
...
    "/api/pilot/config": {
      "get": {
        "tags": ["Config"],
        "summary": "Get configuration",
        "operationId": "getConfig",
        "parameters": [
          { "name": "section", "in": "query", "schema": { "type": "string" }, "description": "Return a single config section" },
          { "name": "sections", "in": "query", "schema": { "type": "string" }, "description": "Comma-separated list of section names" }
        ],
        "responses": {
          "200": { "description": "Config object", "content": { "application/json": { "schema": { "type": "object" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to read config", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      },
      "put": {
        "tags": ["Config"],
        "summary": "Update config section",
        "operationId": "updateConfig",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateConfigRequest" } } } },
        "responses": {
          "200": { "description": "Updated", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "400": { "description": "Missing fields", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to update", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
    "/api/pilot/config/credentials": {
      "put": {
        "tags": ["Credentials"],
        "summary": "Save platform credentials",
        "operationId": "saveCredentials",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SaveCredentialsRequest" } } } },
        "responses": {
          "200": { "description": "Saved", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "400": { "description": "Invalid platform or missing fields", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "description": "Failed to save", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "security": [{ "BearerAuth": [] }]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Auth token from dist/server/data/.token or AUTH_TOKEN env var"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"]
      },
...
      "UpdateConfigRequest": {
        "type": "object",
        "properties": {
          "section": { "type": "string", "description": "Config section name (e.g. telegram, model, platforms)" },
          "values": { "type": "object", "description": "Key-value pairs to merge into the section" }
        },
        "required": ["section", "values"]
      },
      "SaveCredentialsRequest": {
        "type": "object",
        "properties": {
          "platform": { "type": "string", "enum": ["telegram", "discord", "slack", "whatsapp", "matrix", "weixin", "wecom", "feishu", "dingtalk"], "description": "Platform name" },
          "values": { "type": "object", "description": "Credential key-value pairs (e.g. { token: '...', 'extra.app_id': '...' })" }
        },
        "required": ["platform", "values"]
      },
...
    }
  }
}